# docs/openapi.yaml
openapi: 3.0.3
info:
  title: URL Shortener
  version: 1.0.0
  description: >
    Production-style URL shortener. Kept in sync with ShortenerController by
    OpenApiContractTest; a new or removed endpoint fails the build until this
    file is updated.
servers:
  - url: http://localhost:8080
paths:
  /shorten:
    post:
      summary: Create a short link
      operationId: shorten
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Same key always yields the same code. Reusing a key with a different body returns 422.
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShortenRequest' }
      responses:
        '201':
          description: Link created
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShortenResponse' }
        '200':
          description: Idempotent replay of an earlier request (created=false, same code)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShortenResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429':
          description: Rate limited
          headers:
            Retry-After: { schema: { type: integer } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /{code}:
    get:
      summary: Redirect to the destination and record the click
      operationId: redirect
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
        - { name: Referer, in: header, required: false, schema: { type: string } }
      responses:
        '302':
          description: Redirect
          headers:
            Location: { schema: { type: string } }
        '404': { $ref: '#/components/responses/Error' }
        '410': { $ref: '#/components/responses/Error' }
  /stats/{code}:
    get:
      summary: Click analytics
      operationId: stats
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkStats' }
        '404': { $ref: '#/components/responses/Error' }
  /health:
    get:
      summary: Liveness probe
      operationId: health
      responses:
        '200':
          description: Alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ShortenRequest:
      type: object
      required: [url]
      properties:
        url: { type: string, maxLength: 2048, example: 'https://example.com/docs' }
        customAlias: { type: string, pattern: '^[A-Za-z0-9_-]{3,32}$' }
        ttlSeconds: { type: integer, format: int64, minimum: 1 }
    ShortenResponse:
      type: object
      properties:
        code: { type: string }
        shortUrl: { type: string }
        created: { type: boolean }
    LinkStats:
      type: object
      properties:
        code: { type: string }
        targetUrl: { type: string }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        totalClicks: { type: integer, format: int64 }
        uniqueVisitors: { type: integer, format: int64 }
        referrers:
          type: object
          additionalProperties: { type: integer, format: int64 }
        recentClicks:
          type: array
          items:
            type: object
            properties:
              clickedAt: { type: string, format: date-time }
              referrer: { type: string }
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }